move keys/ back into keys.rs
